refactor: extract ConfigEditor into config_editor.py (step 1)
Split the 1,769-line ConfigEditor class out of generate.py
(6,802 -> 5,063 lines, -26%) as the first step of the modularization
plan. The class is moved verbatim; behavior is unchanged.

ConfigEditor has zero `global` statements but read four free names from
generate.py's module scope. These are now injected via the constructor so
the module is self-contained (no shared module-globals, no circular import):

  ConfigEditor(parent, global_config, master_duplication_enabled_var,
               no_user_impersonation_var, on_config_saved=None)

  - global_config            -> self.global_config
  - master_duplication_enabled_var / no_user_impersonation_var: the two
    tk.BooleanVars it bidirectionally syncs with the main window, passed by
    reference so .get()/.set() still mutate the shared vars
  - on_config_saved          -> replaces the globals()['update_dashboard']
    fallback after a save (open_config_editor passes update_dashboard)

validate_url() moved into config_editor.py (it was used only by ConfigEditor).
QUESTIONS_FILE_PATH mirrored as a local constant (UI label only).

Verified: py_compile, pyflakes (no undefined names), and a headless xvfb
smoke test that imports the app and constructs ConfigEditor through its real
injection path. CRLF line endings preserved.
